requirements.txt
fastapi
uvicorn[standard]
google-genai
httpx
pydantic
requests
pydantic-settings